Rename ParseNode.js to parseNode.js.
