+++ src/functions/raisebox.js
@@ -2,7 +2,7 @@
 import defineFunction, {ordargument} from "../defineFunction";
 import buildCommon from "../buildCommon";
 import mathMLTree from "../mathMLTree";
-import {assertNodeType} from "../ParseNode";
+import {assertNodeType} from "../parseNode";
 import {calculateSize} from "../units";
 
 import * as mml from "../buildMathML";
